добавил liquibase скрипт для создания бд, по прежнему выполняю рефакторинг кода
