+++ build.gradle
@@ -21,10 +21,11 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     implementation 'javax.servlet:javax.servlet-api:4.0.1'
     implementation group: 'org.springframework.boot', name: 'spring-boot-starter-validation', version: '3.2.1'
-    implementation 'org.springframework:spring-jdbc:6.0.11'
-    implementation 'org.postgresql:postgresql:42.7.2'
+    implementation 'org.postgresql:postgresql:42.7.4'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.1.0'
     implementation 'org.springframework.boot:spring-boot-starter-data-jdbc:3.3.4'
+
+    implementation 'org.liquibase:liquibase-gradle-plugin:1.1.0'
 
     implementation 'org.projectlombok:lombok:1.18.34'
     compileOnly 'org.projectlombok:lombok:1.18.34'
