+++ docker-compose.yaml
@@ -1,10 +1,10 @@
 version: '3'
 services:
-  postgres:
+  coaches_postgres:
     image: 'postgres:latest'
-    container_name: 'coaches_postgres'
+    container_name: 'pg'
     environment:
-      - 'POSTGRES_DB=coaches_Docker'
+      - 'POSTGRES_DB=coaches'
       - 'POSTGRES_USER=postgres'
       - 'POSTGRES_PASSWORD=0000'
     ports:
@@ -15,15 +15,15 @@ services:
 
   app:
     build: /.
-    container_name: 'java-app'
+    container_name: 'coaches_service'
     ports:
       - "8080:8080"
     environment:
-      - 'DB_URL=jdbc:postgresql://postgres:5432/coaches_Docker'
+      - 'DB_URL=jdbc:postgresql://postgres:5432/coaches'
       - 'DB_USERNAME=postgres'
       - 'DB_PASSWORD=0000'
     depends_on:
-      - postgres
+      - coaches_postgres
     networks:
       - backend
 
@@ -32,23 +32,3 @@ networks:
 
 volumes:
   pg-data:
-
-#version: '3'
-#services:
-#  postgres:
-#    image: 'postgres:latest'
-#    container_name: 'coaches_postgres'
-#    environment:
-#      - 'POSTGRES_DB=coaches'
-#      - 'POSTGRES_USER=postgres'
-#      - 'POSTGRES_PASSWORD=0000'
-#
-#  app:
-#    build: /.
-#    container_name: 'java-app'
-#    ports:
-#      - "8080:8080"
-#    environment:
-#      - 'DB_URL=jdbc:jdbc:postgresql://localhost:5432/coaches'
-#      - 'DB_USERNAME=postgres'
-#      - 'DB_PASSWORD=0000'
