+++ src/main/java/Coaches/persistence/models/CoachDto.java
@@ -1,68 +1,79 @@
-package Coaches.persistence.Models;
+package Coaches.persistence.models;
 
 import Coaches.Entity.Coach;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 import java.sql.Timestamp;
+import java.time.LocalDate;
 import java.util.UUID;
 
 /***
- * Объект содержащий минимум информации о тренере
+ * Объект содержащий полную информацию о тренере
  */
 @Builder
-public class CoachMinimalDto {
+public class CoachDto {
 
     /**
      * Уникальный идентификатор тренера
      */
-    @Schema(description = "Идентификатор")
     public UUID id;
 
     /**
      * Фамилия
      */
-    @Schema(description = "Фамилия")
     public String lastname;
 
     /**
      * Имя
      */
-    @Schema(description = "Имя")
     public String firstname;
 
     /**
      * Отчество
      */
-    @Schema(description = "Отчество")
     public String surname;
 
     /**
      * Возраст
      */
-    @Schema(description = "Возраст")
     public int age;
 
     /**
      * Заархивирован ли профиль тренера? true/false
      */
-    @Schema(description = "Заархивирован ли профиль тренера? true/false")
     public boolean isArchived;
 
     /**
      * Время архивации профиля тренера
      */
-    @JsonFormat(pattern = "dd.MM.yyyy")
-    @Schema(description = "Время архивации профиля")
     public Timestamp archivingTime;
 
-    public static CoachMinimalDto toCoachMinimalDto(Coach coach) {
-        return CoachMinimalDto.builder()
+    /**
+     * Тренер в архиве
+     */
+    public String email;
+
+    /**
+     * Тренер в архиве
+     */
+    @JsonFormat(pattern = "dd.MM.yyyy")
+    public LocalDate birthday;
+
+    /**
+     * Тренер в архиве
+     */
+    public String phoneNumber;
+
+    public static CoachDto toCoachDto(Coach coach) {
+        return CoachDto.builder()
                 .lastname(coach.getLastname())
                 .firstname(coach.getFirstname())
                 .surname(coach.getSurname())
                 .age(coach.getAge())
+                .birthday(coach.getBirthday())
+                .phoneNumber(coach.getPhoneNumber())
+                .email(coach.getEmail())
                 .isArchived(coach.isArchived())
                 .archivingTime(coach.getArchivingTime())
                 .build();
