--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.url=.url=${DATABASE_URL}
-spring.datasource.username=${DATABASE_USERNAME}
-spring.datasource.password=password=${DATABASE_PASSWORD}
-
-
